Lawyers in Singapore
Thank you for taking the time to visit our website. At S SKANDARAJAH and CO, we strive to provide you with quality and professional legal services.
Our motto, "Client Centred Legal Services – because we care, we listen and we help" is something that we take very seriously. We work hard every day to make this slogan a reality.
You could lose your job, your freedom, your reputation and possibly even the right to child custody or visitation. When facing a difficult legal matter, from Criminal Charges to Divorce, you need an effective Criminal Lawyer or Divorce Lawyer working for you. Whether in settlement talks or in the courtroom, we at S SKANDARAJAH and CO who are lawyers in Singapore have a reputation for being fair, yet tough advocates for our clients.