probate lawyer
we represented a home maker wife. Her husband, being a vice-president of a foreign bank, was financially able and ready for a long drawn litigation. Our client’s reasonable desires included care and control of the children, maintenance and equitable division of the matrimonial assets which were all contested by the husband. Multiple interlocutory applications were filed by the husband to wear our client and force her into a settlement. Attempts were made to even bring the children out of Singapore whilst custody issues were being negotiated. Despite the husband’s financial strength and aggressiveness, we managed to successfully obtain a reasonable settlement for her. Our client obtained care and control of the children, maintenance for the children and herself and payment of rent, medical and educational expenses.