Cpf Singapore
Probate, Administration and Resealing of Foreign Grants
Much may be at stake with a death in the family. Apart from dealing with the loss of a loved one, certain legalities will need to be addressed for the family to move on. We work to understand the family’s needs at this time of grief and assist them to make the appropriate applications to Court where the value of the estate is more than $50,000, be it an application for Grant of Probate or Letters of Administration (when there is no will). We take pride in being expeditious and try as much as possible to inconvenience the family during the time of bereavement.