Deed poll
We have successfully assisted many couples to adopt local and overseas children. Our satisfied clients often refer us to their friends and family members. We know and understand that high emotions are involved during the adoption process and work to get all legalities completed at the shortest time possible so that the couples can start life afresh with their new babies.
Recently, our clients wishing to adopt a baby from Malaysia was faced with the task of obtaining a Letter of Support from the Malaysian Embassy. As the Letter of Support was in a new updated format, the Malaysian Embassy in Singapore refused to endorse the Letter of Support and referred our clients to their headquarters in Putrajaya, KL. Unable to get the endorsement on the Letter of Support, our clients’ adoption proceedings were put on hold. When we were engaged, we instructed a counter-part in KL and made personal attendances at Putrajaya, KL and successfully managed to get the Letter of Support for our clients. Shortly after, our clients became the proud new parents of a baby girl.