Immigration law firm
It is common knowledge that the children are the worst-hit victims of a broken marriage. Ideally, it is best for parents to come to an agreement on the care and custody of their children when contemplating separation or divorce.
Sadly, many parents are unable to do so presumably due to the stress of a broken marriage. With our experience in dealing with child custody issues, we are able to advise our clients on the best interests of their children and balance them with our client’s needs and reasonable desires.