Barristers in Singapore
Commissioning and Certification
We offer commissioning services and we have an in-house Notary Public for notarization services. We also provide certification of truth for documents.
Others
Our other areas of practice include drafting of agreements, contracts, deed polls, wills and power of attorney. We also offer legal advice and service in matters relating to contractual disputes, tenancy issues, bankruptcy, debt recovery and many more. Please send us an email with your query and we shall be pleased to assist you.